+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log (@egomobile/http-supertest)
 
+## 0.6.0
+
+- add `onStatusUpdate` to [ISetupTestEventListenerOptions interface](https://egomobile.github.io/node-http-supertest/interfaces/ISetupTestEventListenerOptions.html)
+- fix handling of expected HTTP headers
+- `npm update`s
+
 ## 0.5.0
 
 - **BREAKING CHANGE**: modules requires at least [Node 18+](https://nodejs.org/en/blog/announcements/v18-release-announce) and [@egomobile/http-server](https://github.com/egomobile/node-http-server) `^0.64.2` now